feat: cleanup-config-maptests-input-guard-portos-exceptions-002 - Preserve Dynamic Fanout Spawned-By Guard Coverage Without an Exception
